+++ CHANGELOG.md
@@ -5,3 +5,12 @@
 
 ## 0.0.4
 - patch interface errors
+
+## 1.0.0
+- Remove custom `curve` and `reverseCurve`.
+- Improve the interface update in `BoxDecoration` when hot-reload.
+- Remove `FlowPopupDecoration` and replace it with `BoxDecoration`.
+- Change the name from `childDecoration` to `boxDecoration`.
+- Change the name from `popDecoration` to `popupDecoration`.
+- Change the name from `popBuilder` to `popupBuilder`.
+- Change the name from `childPadding` to `boxPadding` and remove popPadding.